+++ kinto/core/storage/postgresql/__init__.py
@@ -587,31 +587,24 @@ class Storage(StorageBase, MigratorMixin):
                 modified_field=DEFAULT_MODIFIED_FIELD,
                 deleted_field=DEFAULT_DELETED_FIELD,
                 auth=None):
-        query = """
-        WITH collection_filtered AS (
-            SELECT id, last_modified, data, deleted
-              FROM records
-             WHERE {parent_id_filter}
-               AND collection_id = :collection_id
-               {conditions_deleted}
-               {conditions_filter}
-        ),
-        total_filtered AS (
-            SELECT COUNT(id) AS count_total
-              FROM collection_filtered
-             WHERE NOT deleted
-        ),
-        paginated_records AS (
-            SELECT DISTINCT id
-              FROM collection_filtered
-              {pagination_rules}
-        )
-         SELECT count_total,
-               a.id, as_epoch(a.last_modified) AS last_modified, a.data
-          FROM paginated_records AS p JOIN collection_filtered AS a ON (a.id = p.id),
-               total_filtered
-          {sorting}
-         LIMIT :pagination_limit;
+        count_query = """
+        SELECT COUNT(id) AS count_total
+          FROM records
+         WHERE {parent_id_filter}
+           AND collection_id = :collection_id
+           AND NOT deleted
+           {conditions_filter};
+        """
+        select_query = """
+        SELECT id, as_epoch(last_modified) AS last_modified, data
+          FROM records
+         WHERE {pagination_rules}
+           {parent_id_filter}
+           AND collection_id = :collection_id
+           {conditions_deleted}
+           {conditions_filter}
+           {sorting}
+          LIMIT :pagination_limit;
         """
 
         # Unsafe strings escaped by PostgreSQL
@@ -647,7 +640,7 @@ class Storage(StorageBase, MigratorMixin):
         if pagination_rules:
             sql, holders = self._format_pagination(pagination_rules, id_field,
                                                    modified_field)
-            safeholders['pagination_rules'] = 'WHERE {}'.format(sql)
+            safeholders['pagination_rules'] = '{} AND'.format(sql)
             placeholders.update(**holders)
 
         # Limit the number of results (pagination).
@@ -655,13 +648,14 @@ class Storage(StorageBase, MigratorMixin):
         placeholders['pagination_limit'] = limit
 
         with self.client.connect(readonly=True) as conn:
-            result = conn.execute(query.format_map(safeholders), placeholders)
+            result = conn.execute(select_query.format_map(safeholders), placeholders)
             retrieved = result.fetchmany(self._max_fetch_size)
 
-        if len(retrieved) == 0:
-            return [], 0
+            if len(retrieved) == 0:
+                return [], 0
 
-        count_total = retrieved[0]['count_total']
+            result_count = conn.execute(count_query.format_map(safeholders), placeholders)
+            count_total, = result_count.fetchone()
 
         records = []
         for result in retrieved: